fix(eventstream-serde-universal): mark test code for exclusion
