+++ .changeset/wet-balloons-repair.md
@@ -1,0 +1,5 @@
+---
+"@smithy/eventstream-serde-universal": patch
+---
+
+exclude test code from artifact
